requirements.txt
flask~=1.1.2
flask-restful
flask-sqlalchemy
flask-migrate
flask-script
flask_wtf
WTForms~=2.3.1
requests
flask-api
gunicorn
SQLAlchemy~=1.3.17
alembic~=1.4.2
flask_jwt_extended
PyJWT~=1.7.1
flask-bootstrap
alchemy-mock